+++ src/Invoke-Link_function.ps1
@@ -17,19 +17,30 @@
         - Specify the absolute file path in the text file as possible.
           Or Note that when specifying a relative path, the root is the
           location of the current process
+    - If a directory is specified, the names and first lines of the
+      files in that hierarchy are listed.
+        - Collect files recursively with -Recurse option
 
-    Multiple links(lines) in a file available.
-    Lines that empty or beginning with "#" are skipped.
+    Link file settings:
 
-    The link execution app can be any command if -Command option is
-    specified.
-
-    Links written in a text file may or may not be enclosed in
-    single/double quotes.
-
-    If -l or -Location specified, open the file location in explorer
-
-    Environment variables such as ${HOME} can be used for path strings.
+    - Multiple links(lines) in a file available.
+    - Tag
+        - To add tags, add a space + "#tag" to a comment line
+          starting with "#" or "Tag:"
+            - e.g. # commnent #tag-1 #tag-2
+            - e.g. Tag: #tag-1 #tag-2
+        - If you specify a directory as an argument, tags will be output.
+          This is useful when searching linked files by tag.
+    - Skip line
+        - Lines that empty or beginning with "#" are skipped.
+        - Lines that empty or beginning with "Tag:" are skipped.
+    - The link execution app can be any command if -Command option is
+      specified.
+    - Links written in a text file may or may not be enclosed in
+      single/double quotes.
+    - If -l or -Location specified, open the file location in explorer
+      (do not run link)
+    - Environment variables such as ${HOME} can be used for path strings.
 
     Usage:
         i                  ... Equivalent to Invoke-Item .
@@ -37,8 +48,15 @@
         i <file>           ... Invoke-Item <links-writtein-in-text-file>
         i <file> <command> ... command <links-writtein-in-text-file>
         i <file> -l or -Location ... Open <link> location in explorer
-        i <file> -q or -DryRun   ... DryRun (listup links)
+        i <file> -d or -DryRun   ... DryRun (listup links)
         i <file> -e or -Edit     ... Edit <linkfile> using text editor
+    
+    Example of link file with tag:
+        cat ./work/apps/chrome.txt
+
+            # title of link file #app #browser
+            Tag: #hoge #fuga
+            "C:\Program Files\Google\Chrome\Application\chrome.exe"
 
     Input:
         cat ./link/about_Invoke-Item.txt
@@ -65,7 +83,7 @@
     i <file> <command> ... command <links-writtein-in-text-file>
     i <file> <command> -b    ... run command in background
     i <file> -l or -Location ... Open <link> location in explorer
-    i <file> -q or -DryRun   ... DryRun (listup links)
+    i <file> -d or -DryRun   ... DryRun (listup links)
     i <file> -e or -Edit     ... Edit <linkfile> using text editor
 
 .EXAMPLE
@@ -99,8 +117,8 @@
     ## Specify path containing wildcards
     i ./link/a.*
     
-    ## Filee Recursive search
-    i ./work/google-* -Recurse
+    ## Directory recursive search
+    i ./work/ -Recurse
 
 .EXAMPLE
     ## execute if *.ps1 file specified
@@ -121,6 +139,21 @@
     MSRC Security Update Guide 2023-09-15 Chromium: CVE-2023-4903...
     MSRC Security Update Guide 2023-09-15 Chromium: CVE-2023-4904...
     MSRC Security Update Guide 2023-09-15 Chromium: CVE-2023-4905...
+
+.EXAMPLE
+    # tag search
+    
+    ## link file
+    cat ./work/apps/chrome.txt
+        # chrome #app #browser
+        Tag: #hoge #fuga
+        "C:\Program Files\Google\Chrome\Application\chrome.exe"
+
+    ## search by tag
+    i ./work/apps/ | ? tag -match hoge
+        Id Tag                   Name               Line
+        -- ---                   ----               ----
+         1 app,browser,hoge,fuga ./work/apps/chrome # chrome #app #browser
 
 .LINK
     linkcheck
@@ -164,8 +197,22 @@ function Invoke-Link {
         [switch] $AsFileObject,
         
         [Parameter( Mandatory=$False )]
+        [switch] $Extension,
+        
+        [Parameter( Mandatory=$False )]
+        [Alias('a')]
+        [switch] $AllowBulkInput,
+        
+        [Parameter( Mandatory=$False )]
+        [Alias('t')]
+        [string[]] $Tag,
+        
+        [Parameter( Mandatory=$False )]
         [Alias('i')]
         [int[]] $InvokeById,
+        
+        [Parameter( Mandatory=$False )]
+        [int[]] $Id,
         
         [Parameter( Mandatory=$False )]
         [ValidateSet(
@@ -174,14 +221,22 @@ function Invoke-Link {
         [string] $ErrAction = "Stop",
         
         [Parameter( Mandatory=$False )]
+        [Alias('d')]
+        [switch] $DryRun,
+        
+        [Parameter( Mandatory=$False )]
         [Alias('q')]
-        [switch] $DryRun
+        [switch] $Quiet,
+        
+        [parameter( Mandatory=$False, ValueFromPipeline=$True )]
+        [object[]] $InputObject
     )
     # private functions
     function isCommentOrEmptyLine ( [string] $line ){
         [bool] $coeFlag = $False
         if ( $line -match '^#' )   { $coeFlag = $True }
         if ( $line -match '^\s*$' ){ $coeFlag = $True }
+        if ( $line -match '^Tag:' ){ $coeFlag = $True }
         return $coeFlag
     }
     function isLinkHttp ( [string] $line ){
@@ -191,8 +246,8 @@ function Invoke-Link {
         return $httpFlag
     }
     function isLinkAlive ( [string] $uri ){
+        $origErrActPref = $ErrorActionPreference
         try {
-            $origErrActPref = $ErrorActionPreference
             $ErrorActionPreference = "SilentlyContinue"
             $Response = Invoke-WebRequest -Uri "$uri"
             $ErrorActionPreference = $origErrActPref
@@ -202,6 +257,8 @@ function Invoke-Link {
         } catch {
             $StatusCode = $_.Exception.Response.StatusCode.value__
             return $False
+        } finally {
+            $ErrorActionPreference = $origErrActPref
         }
     }
     function editFile ( [string] $fpath ){
@@ -216,60 +273,181 @@ function Invoke-Link {
         }
         return
     }
+    function getRelativePath ( [string] $LiteralPath ){
+        [String] $res = Resolve-Path -LiteralPath $LiteralPath -Relative
+        if ( $IsWindows ){ [String] $res = $res.Replace('\', '/') }
+        return $res
+    }
+    function getMatchesValue {
+        param (
+            [String] $line,
+            [String] $pattern,
+            [Parameter( Mandatory=$False )]
+            [String[]] $replaceChar
+        )
+        $splatting = @{
+            Pattern       = $pattern
+            CaseSensitive = $False
+            Encoding      = "utf8"
+            SimpleMatch   = $False
+            NotMatch      = $False
+            AllMatches    = $True
+        }
+        [String[]] $retAry = ($line | Select-String @splatting).Matches.Value `
+            | ForEach-Object {
+                [String] $writeLine = "$_".Trim()
+                if ( $replaceChar.Count -gt 0 ){
+                    foreach ( $r in $replaceChar ){
+                        $writeLine = $writeLine.Replace($r, '')
+                    }
+                }
+                Write-Output $writeLine
+            }
+        return $retAry
+    }
+    # set variable
+    $hrefList = New-Object 'System.Collections.Generic.List[System.String]'
+    if ( $Files.Count -gt 0){
+        #pass
+    } else {
+        if ( $input.Count -lt 1 ){
+            Write-Error "No input file." -ErrorAction Stop
+        }
+        if ( $True ){
+            Write-Error "Input via pipeline is not allowed." -ErrorAction Stop
+        }
+        return
+        [string[]] $Files = $input | ForEach-Object {
+            if ( ($_ -is [System.IO.FileInfo]) `
+                -or ($_ -is [System.IO.DirectoryInfo]) `
+                -or ($_ -is [System.IO.FileSystemInfo]) ){
+                Write-Output $_.FullName
+            } else {
+                Write-Output "$_"
+            }
+        }
+    }
+    # test bulk input
+    if ( -not $AllowBulkInput ){
+        $bulkList = New-Object 'System.Collections.Generic.List[System.String]'
+        foreach ( $f in $Files ){
+            # expand wildcard
+            Get-Item -Path $f | ForEach-Object {
+                if ( Test-Path -Path $_.FullName -PathType Leaf){
+                    [String] $resPath = getRelativePath $_.FullName
+                    $bulkList.Add($resPath)
+                }
+            }
+        }
+        [String[]] $bulkFileAry = $bulkList.ToArray()
+        if ( $bulkFileAry.Count -gt 1 ){
+            foreach ( $f in $bulkFileAry ){
+                Write-Output $f
+            }
+            Write-Error "Detect input of 5 or more items. To avoid this error, specify the '-AllowBulkInput' option." -ErrorAction Stop
+        }
+    }
+    # main
     [int] $fileCounter = 0
-    foreach ( $f in $Files){
+    foreach ( $f in $Files ){
         # interpret Paths containing wildcards
         if ( Test-Path -Path $f -PathType Container){
             [string[]] $tmpFiles = Get-Item -Path $f `
                 | Resolve-Path -Relative
         } else {
-            [string[]] $tmpFiles = Get-ChildItem -Path $f -Recurse:$Recurse `
+            [string[]] $tmpFiles = Get-ChildItem -Path $f -Recurse:$Recurse -File `
                 | Resolve-Path -Relative
         }
         # set links
         foreach ( $File in $tmpFiles ){
             # is path directory?
-            if ( -not (isLinkHttp $File) ){
-                if ( Test-Path -LiteralPath $File -PathType Container){
-                    if ( $DryRun ){
-                        Write-Output $File
-                        continue
-                    }
-                    # return file paths
-                    #Invoke-Item -LiteralPath $File
-                    Get-ChildItem -LiteralPath $File `
-                        | Sort-Object {
-                            -join ( [int[]] ($_.Name.ToCharArray()) | ForEach-Object { [System.Convert]::ToString($_, 16)})
-                        } `
-                        | ForEach-Object {
-                            $fileCounter++
-                            if ( $InvokeById.Count -gt 0){
-                                if ($InvokeById.Contains($fileCounter)){
-                                    Write-Output "Invoke-Link: $((Resolve-Path -LiteralPath $_.FullName -Relative).Replace('\','/'))"
-                                    Invoke-Link -Files $_.FullName
-                                }
-                                return
+            if ( Test-Path -LiteralPath $File -PathType Container){
+                if ( $DryRun ){
+                    Write-Output $File
+                    continue
+                }
+                # return file paths
+                #Invoke-Item -LiteralPath $File
+                Get-ChildItem -LiteralPath $File -Recurse:$Recurse -File `
+                    | Sort-Object {
+                        -join ( [int[]] ($_.FullName.ToCharArray()) | ForEach-Object { [System.Convert]::ToString($_, 16)})
+                    } `
+                    | ForEach-Object {
+                        $fileCounter++
+                        if ( $InvokeById.Count -gt 0){
+                            if ($InvokeById.Contains($fileCounter)){
+                                [String] $relPath = getRelativePath $_.FullName
+                                Write-Output "Invoke-Link: $relPath"
+                                Invoke-Link -Files $_.FullName
                             }
-                            if (
-                                (Test-Path -LiteralPath $_.FullName -PathType Container) -or `
-                                ($_.Extension -match '\.lnk$|\.exe$|\.dll$|\.xls|\.doc|\.ppt|\.ps1$') `
-                                ){
-                                $hash = [ordered] @{
-                                    Id   = $fileCounter
-                                    Name = $_.Name -replace '\.[^\.]+$', ''
-                                    Line = $Null
-                                }
+                            return
+                        }
+                        if ( $Id.Count -gt 0){
+                            if ($Id.Contains($fileCounter)){
+                                Get-Item -LiteralPath $_.FullName
+                            }
+                            return
+                        }
+                        # set path
+                        [String] $parentPath = Split-Path -Parent $_
+                        [String] $childPath = Split-Path -Leaf $_
+                        [String] $joinedPath = Join-Path -Path $parentPath -ChildPath $childPath
+                        [String] $relativePath = getRelativePath $joinedPath
+                        # remove extension
+                        if ( -not $Extension -and $_.Name -notmatch '^\.') {
+                            [String] $relativePath = $relativePath -replace '\.[^\.]+$', ''
+                        }
+                        if ( Test-Path -LiteralPath $_.FullName -PathType Container){
+                            continue
+                        } elseif ( $_.Extension -match '\.lnk$|\.exe$|\.dll$|\.xls|\.doc|\.ppt|\.ps1$' ){
+                            if ( $Tag.Count -gt 0 ){
+                                continue
                             } else {
                                 $hash = [ordered] @{
                                     Id   = $fileCounter
-                                    Name = $_.Name -replace '\.[^\.]+$', ''
-                                    Line = Get-Content -Path $_.FullName -TotalCount 1 -Encoding utf-8
+                                    Tag  = $Null
+                                    Name = $relativePath
+                                    Line = $Null
                                 }
                             }
-                            [pscustomobject] $Hash
+                        } else {
+                            # get tag
+                            [String] $pat = ' #[^ ]+'
+                            $splatting = @{
+                                Pattern       = $pat
+                                CaseSensitive = $False
+                                Encoding      = "utf8"
+                                AllMatches    = $True
+                                Path          = $_.FullName
+                            }
+                            #[String[]] $tagAry = getMatchesValue $line ' #[^ ]+|^#[^ ]+'
+                            [String[]] $tagAry = (Select-String @splatting).Matches.Value `
+                                | ForEach-Object { Write-Output $("$_".Trim()) }
+                            if ( $Tag.Count -gt 0 ){
+                                [bool] $isMatchTag = $False
+                                $tagAry | ForEach-Object {
+                                    foreach ( $t in $Tag ){
+                                        if ( $_ -match $t ){
+                                            $isMatchTag = $True
+                                            return
+                                        }
+                                    }
+                                }
+                                if ( -not $isMatchTag ){
+                                    return
+                                }
+                            }
+                            [String] $tagStr = ($tagAry -join ", ").Replace('#', '')
+                            $hash = [ordered] @{
+                                Id   = $fileCounter
+                                Tag  = $tagStr
+                                Name = $relativePath
+                                Line = Get-Content -Path $_.FullName -TotalCount 1 -Encoding utf-8
+                            }
                         }
-                    continue
-                }
+                        [pscustomobject] $Hash
+                    }
+                continue
             }
             # is file exist?
             if ( -not ( Test-Path -LiteralPath $File ) ){
@@ -287,6 +465,10 @@ function Invoke-Link {
             if ( $Edit ){
                 editFile $File
                 continue
+            }
+            # output file name
+            if ( -not $Quiet ){
+                Write-Output $File
             }
             # is windows shortcut?
             [string] $ext = (Get-Item -LiteralPath $File).Extension
@@ -346,10 +528,15 @@ function Invoke-Link {
                         Write-Output $linkLine
                     }
                 } | Sort-Object -Unique
-            if ( $Edit ){ continue }
+            if ( $Edit ){
+                continue
+            }
             if ( -not $linkLines ){
-                if ( Test-Path -LiteralPath $File -PathType Container){ continue }
-                "Could not find links in file: $File" | Write-Error -ErrorAction Stop; continue
+                if ( Test-Path -LiteralPath $File -PathType Container){
+                    continue
+                }
+                "Could not find links in file: $File" | Write-Error -ErrorAction Stop
+                continue
             }
             if ( $DryRun ){
                 Write-Host "$File" -ForegroundColor Green
@@ -363,16 +550,20 @@ function Invoke-Link {
                     }
                 continue
             }
+            foreach ( $href in $linkLines ){
+                $hrefList.Add($href)
+            }
         }
         if ( $Location -and -not $Command ){
             continue
         }
-        foreach ( $hlink in $linkLines ){
+        [String[]] $linkAry = $hrefList.ToArray()
+        foreach ( $href in $linkAry ){
             # execute command
             if ( $Command ){
                 [string] $com = $Command
             } else {
-                if ( isLinkHttp $hlink ){
+                if ( isLinkHttp $href ){
                     [string] $com = "Start-Process -FilePath"
                 } elseif ($AsFileObject) {
                     [string] $com = "Get-Item -LiteralPath"
@@ -380,8 +571,8 @@ function Invoke-Link {
                     [string] $com = "Invoke-Item"
                 }
             }
-            Write-Debug $hlink
-            [string] $com = "$com ""$hlink"""
+            Write-Debug $href
+            [string] $com = "$com ""$href"""
             if ( $DryRun ){
                 if ( $BackGround ){
                     [string] $com = "Start-Job -ScriptBlock { Invoke-Expression -Command $com -ErrorAction $ErrAction }"